09-01-2024

+++ backend/routers/products.js
@@ -2,7 +2,7 @@ const express = require("express");
 const Product = require("../models/product");
 const {connection} = require("../database");
 const router = express.Router();
-
+const authenticateToken = require("../helpers/jwt");
 
 // GET endpoint to retrieve product by id
 router.get('/:id', (req, res) => {
@@ -20,13 +20,21 @@ router.get('/', (req, res) => {
 });
 
 // POST endpoint to add a new product
-router.post('/', (req, res) => {
+router.post('/', authenticateToken, (req, res) => {
+    const isAdmin = req.user.isAdmin;
     const { category_id, product_image, product_name, product_price, product_stock, product_desc } = req.body;
-    Product.addProduct(connection, category_id, product_image, product_name, product_price, product_stock, 
-        product_desc, (result) => {
-        res.json(result);
-    });
-});
+    if(isAdmin == "TRUE"){
+        Product.addProduct(connection, category_id, product_image, product_name, product_price, product_stock, 
+            product_desc, (result) => {
+            res.json(result);
+        });
+    }
+    else{ //return status and message
+        res.status(401).json({
+            status: "error",
+            message: "Unauthorized"
+        });
+    } });
 
 // PUT endpoint to update product by id
 router.put('/:id', (req, res) => {
@@ -39,11 +47,20 @@ router.put('/:id', (req, res) => {
 });
 
 // DELETE endpoint
-router.delete('/:id', (req, res) => {
+router.delete('/:id', authenticateToken, (req, res) => {
     const id = req.params.id;
-    Product.deleteProductById(connection, id, (result) => {
-        res.json(result);
-    });
+    const isAdmin = req.user.isAdmin;
+    if(isAdmin == "TRUE"){
+        Product.deleteProductById(connection, id, (result) => {
+            res.json(result);
+        });
+    }
+    else{
+        res.status(401).json({
+            status: "error",
+            message: "Unauthorized"
+        });
+    }
 });
 
 // GET endpoint to get productCount
